github/ryan-file-converter-app/requirements.txt
flask==2.3.3
werkzeug==2.3.7
jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3
flask-login==0.6.2
flask-bcrypt==1.0.1
flask-sqlalchemy==2.5.1
sqlalchemy==1.4.31
pypdf2==3.0.1
pdf2image==1.16.3
pdfkit==1.0.0
pdfrw==0.4.0
pillow>=9.5.0
pytesseract==0.3.10
pdf2docx==0.5.6
fitz==0.0.1.dev2
img2pdf==0.4.4
pdfminer-six==20221105
fpdf==1.7.2
reportlab==4.3.1
soundfile==0.12.1
numpy>=1.24.0
ffmpeg-python==0.2.0
requests==2.28.2
tqdm==4.65.0
uuid==1.30
python-dotenv==1.0.0
gunicorn==20.1.0
babel==2.11.0
python-magic==0.4.27
pydub==0.25.1
# Dev dependencies
pytest==7.2.2
coverage==7.2.1